+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Dependabot: Bump decode-uri-component from 0.2.0 to 0.2.2
+
 ## [0.2.0] - 2022-11-28
 
 ### Changed