+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Add `disabled` property to the `Input` component.
+### Added
+
+- `disabled` property to the `Input` component.
 
 ### Changed